requirements.txt
Django
djoser
psycopg2
requests
djangorestframework
uwsgi
django-cors-headers
django-redis
celery
djangorestframework-simplejwt
django-model-utils
django-better-admin-arrayfield
django-filter
Pillow
django-versatileimagefield
